+++ src/directives/dndDraggable.js
@@ -152,7 +152,6 @@ function ($timeout, $parse, $http, $compile, $q, $templateCache, EventEmitter) {
 
             updatePosition: function () {
                 var position = this.api.getRelBorderedAxis(this.borderOffset).plus(this._offset);
-                console.log(this.api.getRelBorderedAxis())
 
                 wrapper.dndCss(position.getAsCss());
             },
@@ -173,7 +172,7 @@ function ($timeout, $parse, $http, $compile, $q, $templateCache, EventEmitter) {
 
         var defaults = {
             layer: 'common',
-            useAsPoint: false,
+            useAsPoint: true,
             helper: null,
             handle: ''
         };